requirements.txt
attrs==25.3.0
autopep8==2.3.2
aws-cdk-asset-awscli-v1==2.2.242
aws-cdk-asset-node-proxy-agent-v6==2.1.0
aws-cdk-cloud-assembly-schema==44.9.0
aws-cdk-lib==2.203.0
aws-cdk-region-info==1.204.0
cattrs==24.1.3
constructs==10.4.2
dotenv==0.9.9
importlib-resources==6.5.2
jsii==1.112.0
pip==25.1.1
publication==0.0.3
pycodestyle==2.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
six==1.17.0
typeguard==2.13.3
typing-extensions==4.14.0
